refactor: extract metadata accordion
